requirements.txt
python-telegram-bot
python-dotenv
rapidfuzz
groq